Use InnoSetup to create installer
The installer will copy the built dlls into the 3DSMax Folder, and the rest of the required files in pf/KDAB/Kuesa 3DSMax Plugins.
The plugins are registered for the user running the installer. Other users need to follow instructions showed at the
end of the install process and stored in a file in the install folder.
Un-installings seems to leave an empty Kuesa menu in 3DSMax, not sure how to fix.